requirements.txt
clean-text==0.6.0
faiss-cpu==1.8.0.post1
fastparquet==2024.11.0
ipykernel==6.29.5
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langchain-together==0.1.5
langdetect==1.0.9
pandas==2.2.2
pint==0.24.4
PyYAML==6.0.2
PyMuPDF==1.24.9
pymupdf4llm==0.0.12
rank-bm25==0.2.2
requests==2.32.3
sentence-transformers==3.1.0
setuptools==72.1.0
together==1.3.3
torch==2.5.0
unidecode==1.3.8
